requirements.txt
astroid==2.3.1
Django==2.2.4
django-countries==5.5
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.5
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
